WIP: refactor delegate button and extract as react component
