+++ src/client/components/StoryFooter/Buttons.js
@@ -6,17 +6,16 @@ import { scroller } from 'react-scroll';
 import { withRouter, Link } from 'react-router-dom';
 import { Input, Icon, Modal, Menu, Dropdown, Button, message  } from 'antd';
 import classNames from 'classnames';
-import SteemConnect from '../../steemConnectAPI';
 import withAuthActions from '../../auth/withAuthActions';
 import { sortVotes } from '../../helpers/sortHelpers';
 import { getUpvotes, getDownvotes } from '../../helpers/voteHelpers';
-import { delegationAmounts } from '../../helpers/constants';
 import Popover from '../Popover';
 import BTooltip from '../BTooltip';
 import CertifiedUlogger from '../CertifiedUlogger';
 import PopoverMenu, { PopoverMenuItem } from '../PopoverMenu/PopoverMenu';
 import ReactionsModal from '../Reactions/ReactionsModal';
 import USDDisplay from '../Utils/USDDisplay';
+import DelegateButton from './DelegateButton';
 import './Buttons.less';
 
 @injectIntl
@@ -72,7 +71,6 @@ export default class Buttons extends React.Component {
       shareModalLoading: false,
       reactionsModalVisible: false,
       loadingEdit: false,
-      delegateSpMenuVisible: false,
     };
 
     this.handleLikeClick = this.handleLikeClick.bind(this);
@@ -239,20 +237,6 @@ export default class Buttons extends React.Component {
     );
   }
 
-  handleInputCustomSP = (e) => {
-    this.setState({ customSP: e.target.value });
-  }
-
-  handleDelegateSpMenuClick = (e) => {
-    if (e.key !== 'custom') {
-      this.setState({ delegateSpMenuVisible: false });
-    }
-  }
-
-  handleDelegateSpVisibleChange = (flag) => {
-    this.setState({ delegateSpMenuVisible: flag });
-  }
-
   render() {
     const { intl, post, postState, pendingLike, ownPost, defaultVotePercent } = this.props;
 
@@ -317,51 +301,6 @@ export default class Buttons extends React.Component {
       );
     }
 
-    const delegateSpMenu = (
-      <Menu
-        onClick={this.handleDelegateSpMenuClick}
-        style={{ width: '65%' }}
-      >
-        {delegationAmounts.map(
-          amount => {
-            const delegateQuery = {
-              delegatee: post.author,
-              vesting_shares: `${amount} SP`,
-            };
-            return (
-              <Menu.Item key={amount}>
-                <Link
-                  target='_blank'
-                  to={SteemConnect.sign('delegateVestingShares', delegateQuery)}
-                >
-                  <Icon type='rocket' /> {amount} SP
-                </Link>
-              </Menu.Item>
-            );
-          }
-        )}
-        <Menu.Item key={'custom'}>
-          <Input.Group compact>
-            <Input
-              placeholder="Custom"
-              type={'number'}
-              onChange={this.handleInputCustomSP}
-              addonAfter={
-                <span>
-                SP <Button
-                  type='primary'
-                  shape='circle'
-                  icon='rocket' 
-                  target='_blank'
-                  href={SteemConnect.sign('delegateVestingShares', {delegatee: post.author, vesting_shares: `${this.state.customSP} SP`,})}
-                  />
-                </span>
-              }/>
-          </Input.Group>
-        </Menu.Item>
-      </Menu>
-    );
-
     return (
       <div className="Buttons">
         <BTooltip title={likeTooltip}>
@@ -418,19 +357,7 @@ export default class Buttons extends React.Component {
         {postState.isCertifiedUlogger && (
           <span>
             <CertifiedUlogger />
-            <Dropdown
-              overlay={delegateSpMenu}
-              trigger={['click']}
-              onVisibleChange={this.handleDelegateSpVisibleChange}
-              visible={this.state.delegateSpMenuVisible}
-            >
-              <Button
-                size={'small'}
-                style={{ marginLeft: 8, fontSize: 12 }}
-                type={'primary'}>
-                Delegate <Icon type="down" />
-              </Button>
-            </Dropdown>
+            <DelegateButton post={post} />
           </span>
         )}
         {!postState.isReblogged && (
